\name{odseq-package}
\alias{odseq-package}
\alias{odseq}
\docType{package}
\title{
\packageTitle{odseq}
}
\description{
\packageDescription{odseq}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{odseq}
\packageIndices{odseq}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{odseq}

Maintainer: \packageMaintainer{odseq}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
